JGG comments (part 2)
I have now added comments about all the tests.  What is remaining is a
way to prioritize the tests.  That is difficulty because really
everything listed needs to be tested and hopefully should be easy to
create for the UI tests using the Selenium IDE.  I don't know that the
format of this file lends itself well to the prioritization scheme,
either.